pyproject.toml
[project]
name = "chatapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.7.2",
    "langchain==0.2.16",
    "langchain-google-genai>=1.0.10",
    "streamlit>=1.52.1",
    "python-dotenv",
]
